+++ packages/components/psammead-most-read/src/List/index.jsx
@@ -8,7 +8,7 @@ import {
 import Grid from '@bbc/psammead-grid';
 import { mostReadListGridProps } from '../utilities/gridProps';
 
-const StyledGridBase = styled(Grid).attrs({
+const TwoColumnGrid = styled(Grid).attrs({
   role: 'list',
 })`
   list-style-type: none;
@@ -27,7 +27,7 @@ const StyledGridBase = styled(Grid).attrs({
   }
 `;
 
-const StyledGridExtended = styled(StyledGridBase)`
+const MultiColumnGrid = styled(TwoColumnGrid)`
   @media (min-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN}) {
     grid-auto-flow: row;
     grid-template-rows: repeat(
@@ -37,26 +37,19 @@ const StyledGridExtended = styled(StyledGridBase)`
   }
 `;
 
-const MostReadList = ({ numberOfItems, dir, maxTwoColumns, children }) =>
-  maxTwoColumns ? (
-    <StyledGridBase
+const MostReadList = ({ numberOfItems, dir, maxTwoColumns, children }) => {
+  const MostReadListGrid = maxTwoColumns ? TwoColumnGrid : MultiColumnGrid;
+  return (
+    <MostReadListGrid
       {...mostReadListGridProps(maxTwoColumns)}
       dir={dir}
       numberOfItems={numberOfItems}
       forwardedAs="ol"
     >
       {children}
-    </StyledGridBase>
-  ) : (
-    <StyledGridExtended
-      {...mostReadListGridProps(maxTwoColumns)}
-      dir={dir}
-      numberOfItems={numberOfItems}
-      forwardedAs="ol"
-    >
-      {children}
-    </StyledGridExtended>
+    </MostReadListGrid>
   );
+};
 
 MostReadList.propTypes = {
   children: node.isRequired,