+++ packages/components/psammead-most-read/src/Rank/index.jsx
@@ -44,7 +44,7 @@ const doubleDigitWidth = ({ service }) => {
     : doubleDigitDefault;
 };
 
-const StyledWrapperBase = styled.div`
+const TwoColumnWrapper = styled.div`
   @media (max-width: ${GEL_GROUP_0_SCREEN_WIDTH_MAX}) {
     min-width: ${props =>
       listHasDoubleDigits(props) ? doubleDigitWidth(props).group0 : 'auto'};
@@ -78,7 +78,7 @@ const StyledWrapperBase = styled.div`
   }
 `;
 
-const StyledWrapperExtended = styled(StyledWrapperBase)`
+const MultiColumnWrapper = styled(TwoColumnWrapper)`
   @media (min-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN}) {
     min-width: ${props =>
       props.listIndex === 5 && listHasDoubleDigits(props)
@@ -121,8 +121,10 @@ const MostReadRank = ({
 }) => {
   const numerals = serviceNumerals(service);
   const rank = numerals[listIndex];
-  return maxTwoColumns ? (
-    <StyledWrapperBase
+  const RankWrapper = maxTwoColumns ? TwoColumnWrapper : MultiColumnWrapper;
+
+  return (
+    <RankWrapper
       listIndex={listIndex}
       service={service}
       numberOfItems={numberOfItems}
@@ -131,18 +133,7 @@ const MostReadRank = ({
       <StyledSpan service={service} script={script}>
         {rank}
       </StyledSpan>
-    </StyledWrapperBase>
-  ) : (
-    <StyledWrapperExtended
-      listIndex={listIndex}
-      service={service}
-      numberOfItems={numberOfItems}
-      dir={dir}
-    >
-      <StyledSpan service={service} script={script}>
-        {rank}
-      </StyledSpan>
-    </StyledWrapperExtended>
+    </RankWrapper>
   );
 };