+++ packages/components/psammead-radio-schedule/src/ProgramCard/index.jsx
@@ -29,12 +29,15 @@ import { formatUnixTimestamp } from '@bbc/psammead-timestamp-container/utilities
 const CardWrapper = styled.div`
   padding-top: ${GEL_SPACING};
   background-color: ${C_WHITE};
-  position: relative;
+  display: flex;
+  flex-direction: column;
   border: 0.0625rem solid transparent;
+  height: 100%;
 `;
 
 const TextWrapper = styled.div`
   padding: 0 ${GEL_SPACING};
+  flex-grow: 1;
 `;
 
 const HeadingWrapper = styled.h3`