+++ packages/components/psammead-byline/CHANGELOG.md
@@ -2,5 +2,7 @@
 
 | Version       | Description |
 |---------------|-------------|
+| 1.0.2 | [PR#3135](https://github.com/bbc/psammead/pull/3135) Talos - Bump Dependencies - @bbc/gel-foundations |
+| 1.0.1 | [PR#3129](https://github.com/bbc/psammead/pull/3129) Talos - Bump Dependencies - @bbc/gel-foundations |
 | 1.0.0         | [PR#3117](https://github.com/bbc/psammead/pull/3117) Remove alpha tag. |
 | 0.1.0-alpha.0 | [PR#3061](https://github.com/bbc/psammead/pull/3061) Initial creation of package. |