+++ packages/components/psammead-most-read/CHANGELOG.md
@@ -5,6 +5,9 @@
 |---------|-------------|
 | 1.2.0 | [PR#3120](https://github.com/bbc/psammead/pull/3120) Remove `MostRead` and `MostReadSection` as well as other cleanup |
 | 1.1.0 | [PR#3116](https://github.com/bbc/psammead/pull/3116) Added `maxTwoColumns` prop to conditionally render GEL_GROUP_5_SCREEN_WIDTH_MIN media queries |
+| 1.0.9 | [PR#3135](https://github.com/bbc/psammead/pull/3135) Talos - Bump Dependencies - @bbc/gel-foundations |
+| 1.0.8 | [PR#3129](https://github.com/bbc/psammead/pull/3129) Talos - Bump Dependencies - @bbc/gel-foundations |
+| 1.0.7 | [PR#3121](https://github.com/bbc/psammead/pull/3121) Talos - Bump Dependencies - @bbc/psammead-grid |
 | 1.0.6 | [PR#3102](https://github.com/bbc/psammead/pull/3102) Fix double digit column spacing in MostRead |
 | 1.0.5 | [PR#3082](https://github.com/bbc/psammead/pull/3082) Talos - Bump Dependencies - @bbc/psammead-grid |
 | 1.0.4 | [PR#3071](https://github.com/bbc/psammead/pull/3071) Update README with missing content |