+++ packages/components/psammead-navigation/CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 6.0.4 | [PR#3135](https://github.com/bbc/psammead/pull/3135) Talos - Bump Dependencies - @bbc/gel-foundations |
+| 6.0.3 | [PR#3129](https://github.com/bbc/psammead/pull/3129) Talos - Bump Dependencies - @bbc/gel-foundations |
 | 6.0.2 | [PR#3059](https://github.com/bbc/psammead/pull/3059) Update gradient width on scrollable navigation |
 | 6.0.1 | [PR#3030](https://github.com/bbc/psammead/pull/3030) Talos - Bump Dependencies - @bbc/psammead-styles, @bbc/psammead-assets |
 | 6.0.0 | [PR#3028](https://github.com/bbc/psammead/pull/3028) Bring Navigation out of alpha |